+++ himotoki_split/__init__.py
@@ -2,5 +2,5 @@
 
 from himotoki_split.split import SplitResult, split, split_texts
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 __all__ = ["SplitResult", "split", "split_texts", "__version__"]